feat: dynamic nav and routing with login
