+++ game/.streamlit/config.toml
@@ -1,0 +1,2 @@
+[client]
+showSidebarNavigation = false
